Update Remote OTA

+++ lib/ArduinoNative/src/WString.h
@@ -1,6 +1,6 @@
 /* MIT License
  *
- * Copyright (c) 2019 - 2023 Andreas Merkle <web@blue-andi.de>
+ * Copyright (c) 2023 - 2024 Andreas Merkle <web@blue-andi.de>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -369,7 +369,7 @@ public:
      *
      * @return String length
      */
-    unsigned int length() const
+    size_t length() const
     {
         unsigned int length = 0;
         const char*  ptr    = m_buffer;
